Move some context-related things around
